moved the components from the AddMovie folder to the compoments folder.
