+++ client/src/components/DetailsPage/index.js
@@ -4,8 +4,8 @@ import Divider from '@mui/material/Divider';
 import Paper from '@mui/material/Paper';
 import Stack from '@mui/material/Stack';
 import Box from '@mui/material/Box';
-import LikedToggle from '../../pages/AddMovie/LikedToggle';
-import SeenToggle from '../../pages/AddMovie/SeenToggle';
+import LikedToggle from '../LikedToggle';
+import SeenToggle from '../SeenToggle';
 import TextField from '@mui/material/TextField';
 import Button from '@mui/material/Button';
 import Auth from '../../utils/auth';
